Revert "move-image-pick"
This reverts commit 18a0d566
